requirements.txt
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
python-dotenv==0.21.1
SQLAlchemy==1.4.46
Werkzeug==2.2.3
zipp==3.11.0
